+++ src/components/layoutTest.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+export const LayoutTest = () => {
+  return (
+    <>
+      <Link to="/Test">
+        <button className="bg-black text-white">Test page</button>
+      </Link>
+    </>
+  );
+};
